reflow/testutil: RunReflowTests will not allow shadowed test names
Summary:
Previously, `RunReflowTests` in `reflowtestrunner.go` allowed shadowing of test
names, but this can lead to test failures being masked. In this diff, we ensure
that all test names are unique before running, and if not, the test fails with
an informative error message.

Fix [SYSINFRA-712]

Test Plan: Newly added unit tests pass.

Reviewers: smahadevan, spolakh

Reviewed By: spolakh

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D65081

fbshipit-source-id: 8a555ea