reflow/syntax: fix error handling in map comprehensions
Summary:
This diff fixes a bug that could potentially hide errors when evaluating map
comprehensions.

This example demonstrates the bug; `test.Assert` will raise an error if any
value in `bools` is `false`:
```
val TestMapComprErr = [test.Assert(bools) | (_, bools) <- ["a": [true], "b": [false]]]
```

We would expect this code to fail during evaluation without assigning a
value to `TestMapComprErr`. However, it runs successfully because we
overwrite errors as each entry in the map is processed. With the buggy
code, an error will only be surfaced if the last entry has an error. Map
iteration is not ordered based on the map keys, so despite the `b` entry being
last in the map literal, in this case it is actually evaluated before `a`.

The bug is fixed by checking `err` before evaluating the expression for each
entry. If an error is encountered, we return immediately. We could have also
fixed this bug by taking the last error or accumulating all the errors, but
returning the first error was chosen because it is consistent with how errors
are handled in list comprehensions.

Fix [SYSINFRA-333]

Test Plan: Newly added and existing unit tests pass.

Reviewers: jcharumilind, smahadevan, sbagaria

Reviewed By: jcharumilind

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D65139

fbshipit-source-id: 25b7414