[SYSINFRA-580] Handle marshaling of log objects in the exec
Summary:
Moves marshaling of exec logs into the `Exec.Inspect` call.
Adds a flag SaveLogsToRepo to the Exec
    - If true the RemoteLogs objects for stdout/stderr are marshaled into the repo
    - If false the stdout/stderr logs are marshaled into the repo directly
When the exec is created, SaveLogsToRepo is true if the TaskDB is not nil
Digests for stdout/stderr are returned from the exec, which are now synchronously added to the TaskDB after the task completes.
taskDbWrite/taskDbWriteAsync are now removed.

Test Plan: Unit/Integration tests, Local reflow run.

Reviewers: smahadevan, pboyapalli, spolakh

Reviewed By: smahadevan

Subscribers: pbohannon, smahadevan, dnicolaou, pboyapalli

Differential Revision: https://phabricator.grailbio.com/D61189

fbshipit-source-id: 5053112