[SYSINFRA-865] Create cloudwatch log streams lazily
Summary: `cloudwatchLogs` log streams are now created in cloudwatch when the first log is being published to that stream, instead of immediately when `NewStream` is called. If no logs are ever output to a stream, that stream will never be created in cloudwatch.

Test Plan: Added cloudwatch_test.go. Verified as well via remote integration tests / cloudwatch aws console.

Reviewers: smahadevan, pboyapalli, ghorrell

Reviewed By: pboyapalli, ghorrell

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D66318

fbshipit-source-id: bfaf347