[SYSINFRA-819] reflow/taskdb: Add missing ID-index to taskdb setup
Summary:
Add a missing index `ID-index` which previously got missed because none of our
integration tests exercise it.  This was revealed upon rebasing D66573 (which makes changes
to an integration test which exercises this index) on top of D65950.

We need to up the version number to force the infra config package to treat the addition
of this index as a new 'version' requiring a 'setup'.

Test Plan: See also [[ https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-820 | SYSINFRA-820 ]]

Reviewers: pboyapalli

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D66583

fbshipit-source-id: d20fd5c