[SYSINFRA-908] reflow/pool: Shuffle offers before ordering them.
Summary:
When we have a list of offers to choose and allocate from, we shuffle the list first
so that if we have a large number of equal offers, different (concurrent) callers of allocate
do not end up competing for the same (set of) offers. The current behavior will cause
callers to fail to allocate and conclude that there is no usable offer available.

By shuffling, we cause a different set of equal offers to end up in the heap (constructed by pickN)
for the different (concurrent) callers, so that they are picking potentially different offers.

The effect of this change is easily observed in the expanded test TestAllocateScale.

Also remove the unused `pick` method and its test.

Revision also includes an update to `side_effect_test` where we keep retrying if we don't get
the number of log events we expect (from cloudwatch).

Test Plan:
Unit and integration tests.
See also [[ https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-909 | SYSINFRA-909 ]]

Reviewers: pboyapalli, ghorrell

Reviewed By: pboyapalli, ghorrell

Subscribers: jirgon, dnicolaou, ghorrell

Differential Revision: https://phabricator.grailbio.com/D66761

fbshipit-source-id: 78a9b4e