[SYSINFRA-1107] reflow/ec2cluster: Use default subnet only if AZ is compatible
Summary:
When the cluster config has a `VpcId` specified, we compute a mapping of
availability zone names to subnet IDs, provided the mapping is unambiguous.

Previously, if a mapping for an AZ is not available (due to ambiguity or if not setup for the `VpcId`),
and if the cluster config has a `Subnet` specified, then we use that as the default value.

However, this can cause issues if the az and the default subnet are not compatible.

This revision simply adds a compatibility check and if they don't match, we avoid making a spot request AWS API call.

Test Plan:
Unit and integration tests.
See [[ https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-1108 | SYSINFRA-1108 ]]

Reviewers: pboyapalli, jirgon

Reviewed By: pboyapalli

Subscribers: dnicolaou, jclune

Differential Revision: https://phabricator.grailbio.com/D68383

fbshipit-source-id: 51d5e08