[SYSINFRA-1141] reflow/ec2cluster: Reduce alloc attempt interval to 1m
Summary:
This value was changed from `30s` to `5m` in D55292, to address certain issues.
Those underlying issues have since been addressed in D55303.

During the recent low capacity issues, we often see the cluster shutting down
reflowlets (see attached screenshots) despite there being a need for the same ones.
This is likely due to the combination of this `5m` interval (of how often
we attempt to allocate from the existing pool) and the reduction of max idle
duration of a reflowlet from `10m` to `2m` in D66652.

{F1551210}

{F1551209}

{F1551208}

Test Plan: Unit and integration tests

Reviewers: pboyapalli

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D68562

fbshipit-source-id: daa31d0