github.com/grailbio/reflow: remaining vet cleanup
Summary:
This diff addresses the "go vet" complaints under
github.com/grailbio/reflow that remained after D67609.

Other than a change to the values.Env data structure to remove a
copylock violation, this revision should not be behavior-changing.

[BXDS-1093]

Test Plan:
Existing tests pass if I ensure a docker daemon is running
in the background on this machine.

Reviewers: jcharumilind, smahadevan

Reviewed By: jcharumilind, smahadevan

Subscribers: dnicolaou, pboyapalli, cchang

Differential Revision: https://phabricator.grailbio.com/D68851

fbshipit-source-id: 4e78d2c