[SYSINFRA-974] reflow/taskdb: Minor fixes to taskdb setup
Summary:
When experimenting with creating/updating taskdb using `reflow setup-taskdb`,
I discovered that the indices can sometimes be in an `UPDATING` state due to normal usage.

This change updates the logic of `waitForActiveTable` to only look for the relevant states
of indices where we may not want to perform another update.

Test Plan: N/A

Reviewers: pboyapalli

Reviewed By: pboyapalli

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D69012

fbshipit-source-id: 8f4358c