[SYSINFRA-1216] reflow/ec2cluster: Add support to refresh reflow binary in repository in case of errors
Summary:
When a `reflow` binary or pipeline `evaluator` binary starts an EC2 instance,
the `reflow/reflowlet` binary image is shipped to the repository if it doesn't exist.
However, this is attempted only once in the lifetime of the process.

This change will make the binary image to be shipped to the repository again,
if there was a particular type of error encountered while attempting to install the image.
Repetitive errors will be retried upto `maxTries` (currently set to 10),
after which the instance will be given up on.

Test Plan: Unit tests and custom testing (see [[ https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-1262 | SYSINFRA-1262 ]])

Reviewers: pboyapalli

Reviewed By: pboyapalli

Subscribers: dnicolaou, ghorrell

Differential Revision: https://phabricator.grailbio.com/D69685

fbshipit-source-id: 6c65d13