[SYSINFRA-1266] reflow/sched: Fix flaky test TestTaskErrors
Summary:
Fix flakiness of `TestTaskErrors`.

This test is trying to confirm that an otherwise transient state "lost" is reached.
But "lost" tasks are always retried and sometimes the order of acquiring the mutex to read the state
can cause the transition to "running" to happen before we can confirm that the desired state "lost" was reached.

Test Plan:
Before change, the test will sometimes fail:
```
> cd go/src/github.com/grailbio/reflow/sched && go test -test.count 100 -test.run TaskErrors
...
​--- FAIL: TestTaskErrors (0.00s)
    --- FAIL: TestTaskErrors/transient_error (1.05s)
        scheduler_test.go:563: did not reach desired state: lost, instead task state is: running; error: context deadline exceeded
...
```
After change:
```
> cd go/src/github.com/grailbio/reflow/sched && go test -test.count 100 -test.run TaskErrors
...
PASS
ok  	github.com/grailbio/reflow/sched	10.769s
```

Reviewers: pboyapalli

Reviewed By: pboyapalli

Subscribers: dnicolaou, gvitta

Differential Revision: https://phabricator.grailbio.com/D69743

fbshipit-source-id: f89b477