[SYSINFRA-595] reflow: Refactoring and cleanup
Summary:
Most of the changes here are done to facilitate the eventual move of `tool.Runner`
into the `reflow/runtime` package.  This change cleans up the dependency chain
so that the move itself is easier to manage.

There is no functionality change in this revision.

Test Plan: Unit and Integration tests.

Reviewers: ghorrell

Reviewed By: ghorrell

Subscribers: dnicolaou, pboyapalli

Differential Revision: https://phabricator.grailbio.com/D70851

fbshipit-source-id: 61dc1a9