reflow: include error msg for failed tasks in `reflow info` output
Summary:
This diff updates the `reflow info` tool to show the error message
associated with a failed task. The task's error was already being stored
in TaskDB as a string. In this diff, we instead store a recovered
Reflow Error as serialized JSON. This allows us to better format
the error for display. This change is backwards compatible, so `reflow
info` will still be able to display task errors that were saved in
TaskDB prior to this diff.

Jira: [SYSINFRA-1362](https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-1362)

Test Plan:
Built a reflow binary and tested `reflow info` on run IDs known to have
a failed task. The image below shows four different outputs of `reflow info`:
- task error saved as a string (backwards compatible mode)
- full flag, task error saved as a string (backwards compatible mode)
- task error saved as JSON
- full flag, task error saved as JSON

{F1598495}

Reviewers: smahadevan, ghorrell

Reviewed By: smahadevan

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D71281

fbshipit-source-id: 1e81e64