reflow/ec2cluster: manager respects ctx while blocked on waitc
Summary:
The `waitc` channel is unbuffered, so sending to it blocks until a receiver on
the other side of the channel reads from it. In this case, the reader is a
goroutine of the `m.loop` function in the ec2cluster manager.

During shutdown of the reflow runtime:
- the cluster manager can exit before the scheduler
- scheduler waits on `pending` allocs to be 0 before it exits; this means it
  needs to receive allocs (either successful or error'd) on the `notifyc`
  channel
- allocs are returned on the `notifyc` channel when `s.Cluster.Allocate(...)`
  returns
- `s.Cluster.Allocate(...)` in turn calls `c.manager.Allocate(...)` which
  creates a new waiter and writes it to `waitc`
- since the `m.loop` goroutine is no longer reading from `waitc`,
  `c.manager.Allocate(...)` blocks forever and thus prevents the scheduler from
  exiting.

We discovered this deadlock scenario because it occasionally caused some tests
to timeout; the reflow run it self would complete but the runtime shutdown never
finishes so the test eventually times out.

This diff fixes this issue by placing the send to the `waitc` channel in a
`select` clause along with a receive on `ctx.Done()`. This ensures that
if the `ctx` is canceled, we don't get stuck on the `waitc` send.

This addresses part of the problem, but we may still need to look into
why multiple pending allocation requests were in-flight and getting
starved.

Jira: [SYSINFRA-1512](https://jira.ti-apps.aws.grail.com/browse/SYSINFRA-1512)

Test Plan:
Ran a flaky test that was previously susceptible to this issue. Test completed
200 times without getting stuck, which was not possible before the fix.
```
//reflow/modules:internal/bio_wgs_coverage_test                          PASSED in 168.5s
Stats over 200 runs: max = 168.5s, min = 7.2s, avg = 16.1s, dev = 19.2s
```

Reviewers: jclune, spolakh, smahadevan

Reviewed By: smahadevan

Subscribers: dnicolaou

Differential Revision: https://phabricator.grailbio.com/D73567

fbshipit-source-id: 688c56e