[SYSINFRA-1661] Remove calls to DescribeLogStreams
These calls were identified as the cause of slowness (see the Jira task).
Remove these calls as they are no longer needed (since we ensure that
a new log stream is only created once, and there's also a fallback
mechanism which ignores errors encountered when trying to create the same
logstream again).
This change also addresses the issue of high CPU usage identified in D72420.

Unit and integration tests.
Also, before the change, the following command would often take many (often 20+) minutes to run.
But with this change, it runs under or in about a minute.
```
> cd reflow/modules/internal/
> reflow -log debug -cache off run -local bio_gadtools_test.rf
```

fbshipit-source-id: 66c9d39