reflow: Add metrics to compute cost overheads
Add the following metrics which can be used to get a sense of
the reflow cluster's cost overhead.
- Track currently inflight task resource needs and "size"
- Track currently live alloc resources and "size"
- Track the total amount of resources in each reflowlet.

For all the above, "size" is measured as per `resources.ScaledDistance`
which is not an exact cost measurement, but correlates closely with it.

These metrics can be useful to determine the scale of current cost overheads,
measure the impact of any changes made to address them, as well as identify
the level of optimization opportunity that exists.

Approved-by: Jim Clune <jclune@grailbio.com>

GitLab URL: https://gitlab.com/grailbio/grail/-/merge_requests/1306

fbshipit-source-id: a309020